rename: 참가자 관련 테스트 픽스쳐 클래스명 변경
