+++ src/test/java/blackjack/domain/participants/DealerTest.java
@@ -1,10 +1,10 @@
 package blackjack.domain.participants;
 
-import static blackjack.domain.Fixtures.BLACKJACK_WIN_PLAYER;
-import static blackjack.domain.Fixtures.DEALER_17;
-import static blackjack.domain.Fixtures.LOSE_PLAYER;
-import static blackjack.domain.Fixtures.PUSH_PLAYER;
-import static blackjack.domain.Fixtures.WIN_PLAYER;
+import static blackjack.domain.ParticipantFixtures.BLACKJACK_WIN_PLAYER;
+import static blackjack.domain.ParticipantFixtures.DEALER_17;
+import static blackjack.domain.ParticipantFixtures.LOSE_PLAYER;
+import static blackjack.domain.ParticipantFixtures.PUSH_PLAYER;
+import static blackjack.domain.ParticipantFixtures.WIN_PLAYER;
 import static blackjack.domain.card.Denomination.QUEEN;
 import static blackjack.domain.card.Denomination.SEVEN;
 import static blackjack.domain.card.Denomination.SIX;
