+++ src/test/java/blackjack/domain/participants/ParticipantsTest.java
@@ -1,6 +1,6 @@
 package blackjack.domain.participants;
 
-import static blackjack.domain.Fixtures.BETTING_MONEY_1000;
+import static blackjack.domain.ParticipantFixtures.BETTING_MONEY_1000;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.util.List;
